refactor: split home page sections
